Refactor and removed testnet.txos
